Added "new version" check, fixed "illegal characters" while saving mod presets, added the same pwad extensions to the iwad selection , added top menu
Added new version check,  with a form to display the changelog, the local version, the latest version aviable and a button to download it. The check will be executed at startup and open the relative form if a newer version is aviable (can be changed with a checkbox within the form itself). Version check can also be done manually using the new top menu. Current local version is saved in the static class "SharedVar" along with the url to retrive the latest version. The top menu also allows to open the about page (TODO) and quit.